+++ buildVars.py
@@ -25,10 +25,14 @@ addon_info = AddonInfo(
 	# Translators: Long description to be shown for this add-on on add-on information from add-on store
 	addon_description=_("""SIRA (an acronym in Portuguese for Sistema Interno de Ramais e Anotações) is an add-on designed to improve internal communication and the recording of essential information within the Health Department."""),
 	# version
-	addon_version="2025.2.1",
+	addon_version="2025.3.1",
 	# Brief changelog for this version
 	# Translators: what's new content for the add-on version to be shown in the add-on store
-	addon_changelog=_("""Add update check script and update translations"""),
+	addon_changelog=_("""Moves database configuration logic to a new dbConfig.py module and refactors usage across the add-on for safer access.
+
+Updates the settings panel to support new options and improves configuration initialization.
+
+Adds Markdown table support in buildVars.py and improves uninstall logic to optionally clear configuration."""),
 	# Author(s)
 	addon_author="Edilberto Fonseca <edilberto.fonseca@outlook.com>",
 	# URL for the add-on documentation support
@@ -84,7 +88,7 @@ baseLanguage: str = "en"
 # If you need to add support for markup such as tables, fill out the below list.
 # Extensions string must be of the form "markdown.extensions.extensionName"
 # e.g. "markdown.extensions.tables" to add tables.
-markdownExtensions: list[str] = []
+markdownExtensions: list[str] = ["tables"]
 
 # Custom braille translation tables
 # If your add-on includes custom braille tables (most will not), fill out this dictionary.